// src/types/gql.types.ts
export enum OrderDirection {
  asc = "asc",
  desc = "desc"
}

export enum MenuOrderFelid {
  id = "id",
  createdAt = "createdAt",
  updatedAt = "updatedAt",
  creator = "creator",
  creatorName = "creatorName",
  updater = "updater",
  updaterName = "updaterName",
  sort = "sort",
  state = "state",
  name = "name",
  description = "description",
  route = "route",
  icon = "icon",
  title = "title",
  type = "type",
  component = "component",
  outside = "outside",
  isHidden = "isHidden",
  parentId = "parentId"
}
// ...
export interface Department {
  id?: Nullable<string>;
  createdAt?: Nullable<DateTime>;
  updatedAt?: Nullable<DateTime>;
  creator?: Nullable<string>;
  creatorName?: Nullable<string>;
  updater?: Nullable<string>;
  updaterName?: Nullable<string>;
  sort?: Nullable<number>;
  state?: Nullable<number>;
  name?: Nullable<string>;
  description?: Nullable<string>;
  organId?: Nullable<string>;
  parentId?: Nullable<string>;
  Organ?: Nullable<Organ>;
  parent?: Nullable<Department>;
  children?: Nullable<Nullable<Department>[]>;
}
// ...
export interface Dictionary {
  id?: Nullable<string>;
  createdAt?: Nullable<DateTime>;
  updatedAt?: Nullable<DateTime>;
  creator?: Nullable<string>;
  creatorName?: Nullable<string>;
  updater?: Nullable<string>;
  updaterName?: Nullable<string>;
  sort?: Nullable<number>;
  state?: Nullable<number>;
  dictName?: Nullable<string>;
  name?: Nullable<string>;
  description?: Nullable<string>;
  dictEntries?: Nullable<Nullable<DictEntry>[]>;
  Organ?: Nullable<Nullable<Organ>[]>;
}

export interface DictionaryEdge {
  cursor: string;
  node: Dictionary;
}

export interface DictionaryConnection {
  edges?: Nullable<Nullable<DictionaryEdge>[]>;
  pageInfo: PageInfo;
  totalCount: number;
}

export interface PageInfo {
  endCursor?: Nullable<string>;
  hasNextPage: boolean;
  hasPreviousPage: boolean;
  startCursor?: Nullable<string>;
}

export interface Menu {
  id?: Nullable<string>;
  createdAt?: Nullable<DateTime>;
  updatedAt?: Nullable<DateTime>;
  creator?: Nullable<string>;
  updater?: Nullable<string>;
  sort?: Nullable<number>;
  state?: Nullable<number>;
  name?: Nullable<string>;
  description?: Nullable<string>;
  route?: Nullable<string>;
  icon?: Nullable<string>;
  title?: Nullable<string>;
  type?: Nullable<number>;
  component?: Nullable<string>;
  outside?: Nullable<boolean>;
  isHidden?: Nullable<boolean>;
  isCache?: Nullable<boolean>;
  parentId?: Nullable<string>;
  parent?: Nullable<Menu>;
  children?: Nullable<Nullable<Menu>[]>;
  roles?: Nullable<Nullable<Role>[]>;
}

export interface MenuEdge {
  cursor: string;
  node: Menu;
}

export interface MenuConnection {
  edges?: Nullable<Nullable<MenuEdge>[]>;
  pageInfo: PageInfo;
  totalCount: number;
}

export interface Organ {
  id?: Nullable<string>;
  createdAt?: Nullable<DateTime>;
  updatedAt?: Nullable<DateTime>;
  creator?: Nullable<string>;
  creatorName?: Nullable<string>;
  updater?: Nullable<string>;
  updaterName?: Nullable<string>;
  sort?: Nullable<number>;
  state?: Nullable<number>;
  name?: Nullable<string>;
  description?: Nullable<string>;
  uniqueName?: Nullable<string>;
  departments?: Nullable<Nullable<Department>[]>;
  posts?: Nullable<Nullable<Post>[]>;
  roles?: Nullable<Nullable<Role>[]>;
  Dictionary?: Nullable<Nullable<Dictionary>[]>;
  users?: Nullable<Nullable<User>[]>;
}

export interface OrganConnection {
  edges?: Nullable<OrganEdge[]>;
  pageInfo: PageInfo;
  totalCount: number;
}

export interface OrganEdge {
  cursor: string;
  node: Organ;
}

export interface Post {
  id?: Nullable<string>;
  createdAt?: Nullable<DateTime>;
  updatedAt?: Nullable<DateTime>;
  creator?: Nullable<string>;
  updater?: Nullable<string>;
  sort?: Nullable<number>;
  state?: Nullable<number>;
  name?: Nullable<string>;
  description?: Nullable<string>;
  organId?: Nullable<string>;
  creatorName?: Nullable<string>;
  updaterName?: Nullable<string>;
  Organ?: Nullable<Organ>;
  users?: Nullable<Nullable<User>[]>;
}

export interface PostConnection {
  edges?: Nullable<PostEdge[]>;
  pageInfo: PageInfo;
  totalCount: number;
}

export interface PostEdge {
  cursor: string;
  node: Post;
}

export interface Role {
  id?: Nullable<string>;
  createdAt?: Nullable<DateTime>;
  updatedAt?: Nullable<DateTime>;
  creatorName?: Nullable<string>;
  updaterName?: Nullable<string>;
  creator?: Nullable<string>;
  updater?: Nullable<string>;
  sort?: Nullable<number>;
  state?: Nullable<number>;
  name?: Nullable<string>;
  description?: Nullable<string>;
  organId?: Nullable<string>;
  Organ?: Nullable<Organ>;
  users?: Nullable<Nullable<User>[]>;
  menus?: Nullable<Nullable<Menu>[]>;
}

export interface RoleConnection {
  edges?: Nullable<RoleEdge[]>;
  pageInfo: PageInfo;
  totalCount: number;
}

export interface RoleEdge {
  cursor: string;
  node: Role;
}

export interface User {
  id?: Nullable<string>;
  createdAt?: Nullable<DateTime>;
  updatedAt?: Nullable<DateTime>;
  creator?: Nullable<string>;
  updater?: Nullable<string>;
  username?: Nullable<string>;
  password?: Nullable<string>;
  email?: Nullable<string>;
  nickname?: Nullable<string>;
  phone?: Nullable<string>;
  age?: Nullable<number>;
  sex?: Nullable<number>;
  admin?: Nullable<number>;
  avatar?: Nullable<string>;
  sort?: Nullable<number>;
  state?: Nullable<number>;
  roleId?: Nullable<string>;
  creatorName?: Nullable<string>;
  updaterName?: Nullable<string>;
  Role?: Nullable<Role>;
  organs?: Nullable<Nullable<Organ>[]>;
  posts?: Nullable<Nullable<Post>[]>;
}

export interface UserConnection {
  edges?: Nullable<UserEdge[]>;
  pageInfo: PageInfo;
  totalCount: number;
}

export interface UserEdge {
  cursor: string;
  node: User;
}

export type DateTime = any;
export type JWT = any;
type Nullable<T> = T | null;